# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Haradan API
  version: 0.1.0
  description: |-
    Phase-one Haradan HTTP API contract.

    Source of truth: docs/13 (data model) and docs/14 (use-cases).
    All paths under /api. Versioned business routes use /v1.
    BO operations require canonical role=admin AND session clientContext=ADMIN_BO.
    Client-supplied role/owner/actor values are not trusted.
    Access tokens are short-lived; refresh tokens are rotated and never stored as plaintext.
    Worker/internal use-cases are not exposed as HTTP operations.
servers:
- url: "/api"
  description: API base
...
paths:
...
  "/v1/media/{assetId}/{profile}":
    get:
      operationId: GetPublicMedia
      summary: GetPublicMedia
      description: |-
        Exposure PUBLIC. Actor: Anonymous. Streams a READY delivery profile for
        the asset. Object keys and storage credentials never appear.
        HEAD is registered outside OpenAPI (same handler) because gin-server
        generation for HEAD binary responses is awkward; clients may use HEAD
        for cache validators without a body.
      tags:
      - Media
      x-function-id: MEDIA-PUBLIC-01
      x-exposure: PUBLIC
      x-actor: Anonymous
      responses:
        '200':
          description: Binary image bytes
          headers:
            Content-Type:
              schema:
                type: string
            Content-Length:
              schema:
                type: string
            ETag:
              schema:
                type: string
            Last-Modified:
              schema:
                type: string
            Cache-Control:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
                example: inline
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '304':
          description: Not Modified
        '400':
          "$ref": "#/components/responses/BadRequest"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
        '503':
          "$ref": "#/components/responses/DependencyUnavailable"
      parameters:
      - "$ref": "#/components/parameters/AssetIdPath"
      - name: profile
        in: path
        required: true
        schema:
          "$ref": "#/components/schemas/MediaDeliveryProfile"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '404':
        - NOT_FOUND
        '500':
        - INTERNAL_ERROR
        '503':
        - DEPENDENCY_UNAVAILABLE
...
  "/v1/admin/jobs":
    get:
      operationId: ListAdminJobs
      summary: ListAdminJobs
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Lists scheduled job
        definitions ordered by key ASC. Requires role=admin and session
        clientContext=ADMIN_BO.'
      tags:
      - AdminJobs
      x-function-id: JOB-ADMIN-01
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JobAdminListResponse"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      x-error-codes:
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '500':
        - INTERNAL_ERROR
  "/v1/admin/jobs/{jobId}":
    get:
      operationId: GetAdminJob
      summary: GetAdminJob
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Requires role=admin and
        session clientContext=ADMIN_BO.'
      tags:
      - AdminJobs
      x-function-id: JOB-ADMIN-02
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JobAdminView"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/JobIdPath"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '500':
        - INTERNAL_ERROR
    patch:
      operationId: UpdateAdminJob
      summary: UpdateAdminJob
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. expectedVersion optimistic
        concurrency. Requires role=admin and session clientContext=ADMIN_BO.'
      tags:
      - AdminJobs
      x-function-id: JOB-ADMIN-03
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JobAdminView"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/JobIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateJobRequest"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - STALE_VERSION
        - INVALID_STATE
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
  "/v1/admin/jobs/{jobId}/run":
    post:
      operationId: RunAdminJob
      summary: RunAdminJob
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Enqueues a manual run.
        Optional referenceDate for jobs that support it. Requires role=admin and
        session clientContext=ADMIN_BO.'
      tags:
      - AdminJobs
      x-function-id: JOB-ADMIN-04
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JobRunAcceptedResponse"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '409':
          "$ref": "#/components/responses/Conflict"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/JobIdPath"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/RunJobRequest"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '409':
        - INVALID_STATE
        - CONFLICT
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
  "/v1/admin/jobs/{jobId}/history":
    get:
      operationId: ListAdminJobHistory
      summary: ListAdminJobHistory
      description: 'Exposure BO_AUTH. Actor: admin+ADMIN_BO. Cursor pagination ordered
        by createdAt DESC, id DESC. Requires role=admin and session
        clientContext=ADMIN_BO.'
      tags:
      - AdminJobs
      x-function-id: JOB-ADMIN-05
      x-exposure: BO_AUTH
      x-actor: admin+ADMIN_BO
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JobHistoryPage"
        '400':
          "$ref": "#/components/responses/BadRequest"
        '401':
          "$ref": "#/components/responses/Unauthorized"
        '403':
          "$ref": "#/components/responses/Forbidden"
        '404':
          "$ref": "#/components/responses/NotFound"
        '422':
          "$ref": "#/components/responses/ValidationError"
        '500':
          "$ref": "#/components/responses/InternalError"
      security:
      - BearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/JobIdPath"
      - "$ref": "#/components/parameters/Cursor"
      - "$ref": "#/components/parameters/Limit"
      x-error-codes:
        '400':
        - VALIDATION_ERROR
        '401':
        - UNAUTHENTICATED
        - SESSION_REVOKED
        '403':
        - FORBIDDEN
        - ACCOUNT_INACTIVE
        '404':
        - NOT_FOUND
        '422':
        - VALIDATION_ERROR
        '500':
        - INTERNAL_ERROR
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token. Session/refresh validated server-side.
        Role/owner/actor from client are not trusted.
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor
...
    AssetIdPath:
      name: assetId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    JobIdPath:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
      - key
      - value
      properties:
        key:
          type: string
        value:
          type: string
    FieldValidationError:
      type: object
      additionalProperties: false
      required:
      - field
      - message
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
    ErrorResponse:
      type: object
      required:
      - code
      - message
      - traceId
      additionalProperties: false
      properties:
        code:
          "$ref": "#/components/schemas/DomainErrorCode"
        message:
          type: string
        traceId:
          type: string
          minLength: 1
        details:
          type: array
          items:
            "$ref": "#/components/schemas/ErrorDetail"
        fieldErrors:
          type: array
          items:
            "$ref": "#/components/schemas/FieldValidationError"
...
    DomainErrorCode:
      type: string
      enum:
      - VALIDATION_ERROR
      - UNAUTHENTICATED
      - FORBIDDEN
      - ACCOUNT_INACTIVE
      - EMAIL_NOT_VERIFIED
      - NOT_FOUND
      - CONFLICT
      - STALE_VERSION
      - INVALID_STATE
      - OWNERSHIP_REQUIRED
      - DUPLICATE
      - TOKEN_INVALID
      - TOKEN_EXPIRED
      - TOKEN_ALREADY_USED
      - SESSION_REVOKED
      - REFRESH_REPLAY_DETECTED
      - RATE_LIMITED
      - DEPENDENCY_UNAVAILABLE
      - PROCESSING_NOT_READY
      - QUERY_TOO_COMPLEX
      - INTERNAL_ERROR
...
    MediaDeliveryProfile:
      type: string
      enum:
      - HOMEPAGE
      - DETAIL
      - SEARCH
      - BANNER
...
    JobType:
      type: string
      enum:
      - TJK_SYNC
      - PACKAGE_EXPIRY_SCAN
      - MEDIA_RECONCILE
    JobRunStatus:
      type: string
      enum:
      - QUEUED
      - LEASED
      - SUCCEEDED
      - FAILED
      - CANCELLED
      - DEAD
    JobAdminView:
      type: object
      additionalProperties: false
      required:
      - id
      - key
      - name
      - jobType
      - cronExpression
      - isActive
      - timeoutSeconds
      - supportsReferenceDate
      - version
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        key:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        jobType:
          "$ref": "#/components/schemas/JobType"
        cronExpression:
          type: string
          minLength: 1
        isActive:
          type: boolean
        timeoutSeconds:
          type: integer
          minimum: 1
        supportsReferenceDate:
          type: boolean
        version:
          type: integer
          minimum: 1
        lastRunAt:
          type: string
          format: date-time
          nullable: true
        nextRunAt:
          type: string
          format: date-time
          nullable: true
        lastStatus:
          nullable: true
          allOf:
          - "$ref": "#/components/schemas/JobRunStatus"
        lastDurationMs:
          type: integer
          minimum: 0
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobAdminListResponse:
      type: object
      additionalProperties: false
      required:
      - items
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/JobAdminView"
    UpdateJobRequest:
      type: object
      additionalProperties: false
      required:
      - expectedVersion
      properties:
        expectedVersion:
          type: integer
          minimum: 1
        cronExpression:
          type: string
          minLength: 1
        isActive:
          type: boolean
        timeoutSeconds:
          type: integer
          minimum: 1
    RunJobRequest:
      type: object
      additionalProperties: false
      properties:
        referenceDate:
          type: string
          format: date
          nullable: true
    JobRunAcceptedResponse:
      type: object
      additionalProperties: false
      required:
      - jobId
      - runId
      properties:
        jobId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
    JobHistoryItem:
      type: object
      additionalProperties: false
      required:
      - id
      - jobId
      - status
      - executionType
      - createdAt
      - updatedAt
      properties:
        id:
          type: string
          format: uuid
        jobId:
          type: string
          format: uuid
        status:
          "$ref": "#/components/schemas/JobRunStatus"
        executionType:
          type: string
          enum:
          - SCHEDULED
          - MANUAL
        referenceDate:
          type: string
          format: date
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        durationMs:
          type: integer
          minimum: 0
          nullable: true
        lastError:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobHistoryPage:
      type: object
      additionalProperties: false
      required:
      - items
      - hasMore
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/JobHistoryItem"
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
  responses:
    BadRequest:
      description: Syntactic request or parameter parsing failure (malformed path/query value, UUID/integer/enum/boolean format, or unusable required parameter)
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthenticated or session revoked
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden / inactive / email not verified / ownership / BO context
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict / stale version / invalid state / duplicate
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation or query too complex
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
...
    DependencyUnavailable:
      description: External dependency unavailable
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal error
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/ErrorResponse"